requirements.txt
asgiref==3.8.1
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
cssselect2==0.7.0
Django==5.1
djangorestframework==3.15.2
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
fonttools==4.53.1
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.0.3
html5lib==1.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
locust==2.31.3
MarkupSafe==2.1.5
msgpack==1.0.8
pillow==10.4.0
psutil==6.0.0
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.11.0
pyphen==0.16.0
pyzmq==26.1.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tinycss2==1.3.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.3
zope.event==5.0
zope.interface==7.0.1
zopfli==0.2.3
